requirement.txt
langchain-ollama
psycopg2
langchain-community
langchain-experimental
langchain-groq
python-dotenv